Add default Animal and Monster entity groups to config.yml #195
These were already supported but I also added an icon option.
