+++ src/main/java/world/bentobox/limits/commands/admin/OffsetCommand.java
@@ -7,13 +7,19 @@
 package world.bentobox.limits.commands.admin;
 
 
-import com.google.common.base.Enums;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.stream.Collectors;
+
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
 import org.bukkit.entity.EntityType;
 import org.eclipse.jdt.annotation.Nullable;
-import java.util.*;
-import java.util.stream.Collectors;
+
+import com.google.common.base.Enums;
 
 import world.bentobox.bentobox.api.commands.CompositeCommand;
 import world.bentobox.bentobox.api.localization.TextVariables;
